+++ src/client/app/help/help-routing.module.ts
@@ -1,8 +1,9 @@
 import { NgModule }     from '@angular/core';
 import { RouterModule } from '@angular/router';
 
+import { AuthGuard } from 'app/core';
+
 import { HelpComponent } from './help.component';
-import { AuthGuard } from 'app/core/auth-guard.service';
 
 @NgModule({
 	imports: [
